requirements.txt
autopep8==1.6.0
Django==2.2.9
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.2
Markdown==3.3.6
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
zipp==3.7.0
